do not replace operation with `@override` when `listOperationsInOperationGroup`
